feat: impl float variant
